Fixes #1
Initial implementation of comment support. Every document, section or key may have a string array of comments associated with it, accessed using `getComments()` and `getKeyComments()` and set using `setComments()` and `setKeyComments()`. `INIWriter` has a new `CommentPlacementMode` to adjust where comments are placed when they are written back.

There is still more to do.

  * Support reading of comments
  * Check schema and set facades still work properly
  * Make use of comments in serialization
  * Comment change events